chore: moved util/votecount functions to views/votecount
