+++ src/events/commands/voting/votecount.ts
@@ -1,6 +1,6 @@
-import { calculateVoteCount, formatVoteCount } from '../../../util/votecount';
 import { getVoteCounter } from '@models/votecounter';
 import { SlashCommand } from '@structures/interactions/SlashCommand';
+import { calculateVoteCount, formatVoteCount } from '@views/votecounter';
 
 export default new SlashCommand('votecount')
 	.setDescription('[GAME] View the vote count')
